[PBA-JEST] React-native mock setup
Need to mock react-native module, because if the component being tested requires it, jest currently has problems mocking it. Right now, the only workaround I found is to mock it.

Also, react-native mock is a very basic implementation suited for the needs to test utils.js only. It may need to be improved when testing in more components.