[PBA-2767] Fix timeout usage
setTimeout was being called more than once. We may need to eliminate setTimeout completely, but for now this works. I added React's recommended Timer mixin, that substitute the global setTimeout with their own, so is safer to use on React native.

One thing I don't like with this timeout implementation, is that we get a warning in the console each time the timer is updated: 'Warning: setState(...): Cannot update during an existing state transition (such as within `render`). Render methods should be a pure function of props and state.' We may need to think in another way to update the timer, as they way I'm doing it, I'm updating a var that triggers a re-render of the UI within the render function, and that shouldn't happen.