[PBA-2853] UI changes to icons
* Play button changed a bit of styling
* The whole Up next component is 80 px height, as in the mocks.
* The dismiss (x) button is aligned to the title.

The dismiss button was using RectButton.js to render it, but I'm not using it because it isn't that easy to position in the screen, you can't easily position the RectButton component vertically, and wasn't able to align it with the title. I do was able to align it by modifying the RectButton code, but I didn't go with this approach as this may break other screens that are using the RectButton, so I decide to use my own custom component to render the dismiss button, which is very
simple.