[PBA-2479] Present ActivityViewController on top most UIViewController
When we toggle to fullscreen mode, we delete the link in the controller hierarchy to the OOSkinViewController, as we show the view of it in the window bounds to make it fullscreen. When we do this, and we try to present the social view (ActivityViewController) on any of the existant view controllers, it won't show because it will be below the fullscreen view, and the modal ActivityViewController's view is never visible. To fix this, when going fullscreen, I'm presenting an
empty UIViewController, such that when we try to present the ActivityViewController, it will be able to appear on top of everything.

This is the easier fix without changing anything of the former fullscreen code.