[PBA-3854] remove FontAwesome from react code
bottomOverlay: in style, removed fontawesome, changed color(text only) to backgroundColor, used large border radius so it's not an issue

discoveryPanel, removed warning icon.  It didn't add anything to the experience, and is an error case

Endscreen, we used to use fontawesome for replay button but since have never used it.

MoreOptionsScreenStyles: closeIconStyle was obsolete, unused