The problem was that the overlay's clickable area was larger than the visible overlay.  Fixed the problem by making the visible overlay be the clickable object.
Fixed viewView._renderAdOverlay by changing changing the View that contains the image and the dismiss icon to a TouchableHighlight.  Changed the outer TouchableHighlight to a View.  Moved the onPress() to inside the new TouchableHighlight. Had to wrap the image and dismiss icon in a View because a TouchableHighlight can only have 1 child.