[PBA-5493] Show Discovery on end screen correctly
We weren’t showing the discovery on the end screen with showUpNext = false and screenToShowOnEnd = “discovery”. This was a bug since we would only show discovery on end screen if we dismissed the upNext widget previously. Added a logic operation bypassing the upNextDismissed question if the showUpNext button was set to false, this way we would only care about screenToShowOnEnd.