[PLAYER-1551] Update fullscreen logic for iOS
In our fullscreen logic we presented an empty ViewController that seemed to never be visible. I changed that and set this new ViewController to be the window view controller, meaning it will be on top of everything.

Going out of fullscreen mode we remove the references we set for fullscreen.