[PLAYER-1552] Video player area acts as play/pause button when Accessibility is enabled
We always hide the play/pause button when the screen reader is enabled. Also, if the screen reader is enabled the player view area acts as play/pause button if you double tap on it. This way the play/pause button doesn't interfiere with the video content as it was always in the middle of the screen.